third_party/blink/web_tests/wpt_internal/css/css-flexbox/balance/flex-wrap-computed.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/3070">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<body>
<div id="target"></div>
<script>
test_computed_value("flex-wrap", "nowrap");
test_computed_value("flex-wrap", "wrap");
test_computed_value("flex-wrap", "wrap-reverse");
test_computed_value("flex-wrap", "balance");
test_computed_value("flex-wrap", "wrap balance", "balance");
test_computed_value("flex-wrap", "wrap-reverse balance");
test_computed_value("flex-wrap", "balance wrap", "balance");
test_computed_value("flex-wrap", "balance wrap-reverse", "wrap-reverse balance");
test_computed_value("flex-wrap", "balance 1", "balance");
test_computed_value("flex-wrap", "wrap balance 1", "balance");
test_computed_value("flex-wrap", "wrap-reverse balance 1", "wrap-reverse balance");
test_computed_value("flex-wrap", "balance 2");
test_computed_value("flex-wrap", "wrap balance 2", "balance 2");
test_computed_value("flex-wrap", "wrap-reverse balance 2");
test_computed_value("flex-wrap", "balance 2 wrap", "balance 2");
test_computed_value("flex-wrap", "balance 2 wrap-reverse", "wrap-reverse balance 2");
test_computed_value("flex-wrap", "balance round(up, calc(7/2))", "balance 4");
</script>
</body>
</html>
